requirements.txt
requests
discord-py-slash-command